pyproject.toml
[tool.poetry]
name = "trending-github"
version = "0.1.0"
description = ""
authors=["acalio"] 
[tool.poetry.dependencies]
python = ">3.9"
pdoc3 = "^0.10.0"
notebook = "^6.4.10"
pandas = "^1.4.2"
click = "^8.1.3"
networkx = "^2.8"
tqdm = "^4.64.0"
pipe = "^1.6.0"
openpyxl = "^3.0.10"
hydra-core = "^1.2.0"
PyYAML = "^6.0"
fsspec = "^2022.10.0"
gcsfs = "^2022.10.0"
arrow = "^1.2.3"
sqlalchemy = "^1.4.43"
rich = "^13.0.0"
pyarrow = "^10.0.1"
duckdb-engine = "^0.6.6"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"
pylint = "^2.13.9"
black = "^22.3.0"
jupyter = "^1.0.0"
isort = "^5.10.1"
jupyterlab = "^3.4.3"
jupyterlab-system-monitor = "^0.8.0"
matplotlib = "^3.5.2"
seaborn = "^0.11.2"
interrogate = "^1.5.0"

[tool.poetry.group.dev.dependencies]
ipython-sql = "^0.4.1"
nb-black = "^1.0.7"
jupyter-emacskeys = "^0.2.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build   
)/ 
'''
